refactor: import 부분 public API로 수정
